+++ cmd/keytransparency-client/cmd/root.go
@@ -23,8 +23,8 @@ import (
 	"os"
 	"time"
 
-	"github.com/google/keytransparency/cmd/keytransparency-client/grpcc"
 	"github.com/google/keytransparency/core/authentication"
+	"github.com/google/keytransparency/core/client/grpcc"
 	"github.com/google/keytransparency/core/client/kt"
 
 	"github.com/google/trillian"
